+++ src/LocalStack.Client.Extensions/GlobalUsings.cs
@@ -2,6 +2,7 @@
 global using System.Diagnostics.CodeAnalysis;
 global using System.Reflection;
 global using System.Runtime.Serialization;
+global using System.Text;
 
 global using Microsoft.Extensions.Configuration;
 global using Microsoft.Extensions.DependencyInjection;